set the basic menu page
