+++ .gitignore
@@ -1,7 +1,7 @@
 \#*\#
 *~
 target/
-
+.DS_Store
 # Compiled class file
 *.class
 
@@ -25,3 +25,10 @@ target/
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+tokens/
+
+src/main/resources/credentials.json
+
+.env
+
