+++ pom.xml
@@ -1,6 +1,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-	 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
   <!-- model version - always 4.0.0 for Maven 2.x POMs -->
   <modelVersion>4.0.0</modelVersion>
@@ -9,25 +9,49 @@
     <!-- project coordinates - values which uniquely identify this project -->
 
   <groupId>edu.ucsb.cs56.pconrad</groupId>
-  <artifactId>sparkjava-demo-01</artifactId>
+  <artifactId>spark-mustache-demo-02</artifactId>
   <version>1.0</version>
-  <url>https://ucsb-cs56-pconrad.github.io/sparkjava-01</url>
+  <url>https://ucsb-cs56-pconrad.github.io/sparkjava-mustache-example</url>
   
   <!-- configure what mvn package produces -->
   
   <packaging>jar</packaging>
-  <name>sparkjava-demo-01</name>
+  <name>sparkjava-mustache-example-02</name>
   
   <!-- what version of Java should we use and what encoding for files? -->
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
+    <userDefinedMainClass>edu.ucsb.cs56.pconrad.SparkMustacheDemo02</userDefinedMainClass>
+    <userDefinedHerokuApp>TBD</userDefinedHerokuApp>
   </properties>
 
   <!-- what other software does our project depend on (third-party JARs, etc.) -->
   <dependencies>
 
+    
+    <!-- https://mvnrepository.com/artifact/com.sparkjava/spark-template-mustache -->
+    <dependency>
+      <groupId>com.sparkjava</groupId>
+      <artifactId>spark-template-mustache</artifactId>
+      <version>2.7.1</version>
+    </dependency>
+
+    <!-- log4j -->
+    
+    <dependency>
+       <groupId>org.slf4j</groupId>
+       <artifactId>slf4j-api</artifactId>
+       <version>1.7.5</version>
+   </dependency>
+   <dependency>
+       <groupId>org.slf4j</groupId>
+       <artifactId>slf4j-log4j12</artifactId>
+       <version>1.7.5</version>
+  </dependency>
+
+  
     <!-- JUnit unit testing (https://junit.org/) -->    
     <dependency>
       <groupId>junit</groupId>
@@ -57,11 +81,11 @@
 
     <plugins>
       <plugin>
-	<groupId>org.apache.maven.plugins</groupId>
-	<artifactId>maven-project-info-reports-plugin</artifactId>
-	<version>2.9</version>
-	<!-- SEE: https://stackoverflow.com/questions/28098566/maven-site-warning-the-repository-url-https-maven-repository-dev-java-net-no -->
-	<configuration>
+  <groupId>org.apache.maven.plugins</groupId>
+  <artifactId>maven-project-info-reports-plugin</artifactId>
+  <version>2.9</version>
+  <!-- SEE: https://stackoverflow.com/questions/28098566/maven-site-warning-the-repository-url-https-maven-repository-dev-java-net-no -->
+  <configuration>
           <dependencyLocationsEnabled>false</dependencyLocationsEnabled>
         </configuration>
       </plugin>
@@ -82,9 +106,9 @@
       <!-- See: https://www.baeldung.com/executable-jar-with-maven -->
       
       <plugin>
-	<groupId>org.apache.maven.plugins</groupId>
-	<artifactId>maven-assembly-plugin</artifactId>
-	<executions>
+  <groupId>org.apache.maven.plugins</groupId>
+  <artifactId>maven-assembly-plugin</artifactId>
+  <executions>
           <execution>
             <phase>package</phase>
             <goals>
@@ -93,7 +117,7 @@
             <configuration>
               <archive>
                 <manifest>
-		    <mainClass>edu.ucsb.cs56.pconrad.SparkDemo01</mainClass>
+        <mainClass>${userDefinedMainClass}</mainClass>
                 </manifest>
               </archive>
               <descriptorRefs>
@@ -101,42 +125,42 @@
               </descriptorRefs>
             </configuration>
           </execution>
-	</executions>
+  </executions>
       </plugin>
 
       <plugin>
-	<groupId>org.codehaus.mojo</groupId>
-	<artifactId>exec-maven-plugin</artifactId>
-	<version>1.6.0</version>
-	<executions>
-	  <execution>
-	    <id>my-execution</id>
-	    <goals>
-	      <goal>java</goal>
-	    </goals>
-	  </execution>
-	</executions>
-	<configuration>
-	  <mainClass>edu.ucsb.cs56.pconrad.SparkDemo01</mainClass> 
-	</configuration>
+  <groupId>org.codehaus.mojo</groupId>
+  <artifactId>exec-maven-plugin</artifactId>
+  <version>1.6.0</version>
+  <executions>
+    <execution>
+      <id>my-execution</id>
+      <goals>
+        <goal>java</goal>
+      </goals>
+    </execution>
+  </executions>
+  <configuration>
+    <mainClass>${userDefinedMainClass}</mainClass>
+  </configuration>
       </plugin>
 
       <plugin>
-	<groupId>com.heroku.sdk</groupId>
-	<artifactId>heroku-maven-plugin</artifactId>
-	<version>2.0.3</version>
-	<configuration>
+  <groupId>com.heroku.sdk</groupId>
+  <artifactId>heroku-maven-plugin</artifactId>
+  <version>2.0.3</version>
+  <configuration>
           <jdkVersion>1.8</jdkVersion>
           <!-- Use your own application name -->
-	  <!-- at Heroku CLI, use heroku apps to list, or use Heroku Dashboard -->
-          <appName>ucsb-cs56-zihaozhang-08-28-18</appName> 
+    <!-- at Heroku CLI, use heroku apps to list, or use Heroku Dashboard -->
+          <appName>${userDefinedHerokuApp}</appName>
           <processTypes>
             <!-- Tell Heroku how to launch your application -->
             <!-- You might have to remove the ./ in front   -->
-            <web>java $JAVA_OPTS -jar target/sparkjava-demo-01-1.0-jar-with-dependencies.jar</web>
+            <web>java $JAVA_OPTS -jar target/${project.name}-${project.version}-jar-with-dependencies.jar</web>
 
           </processTypes>
-	</configuration>
+  </configuration>
       </plugin>
