+++ src/main/resources/log4j copy.properties
@@ -1,0 +1,11 @@
+# Root logger option
+log4j.rootLogger=DEBUG, stdout
+
+log4j.logger.SparkMustacheDemo01=DEBUG, stdout
+
+# Redirect log messages to console
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target=System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+
