+++ src/main/resources/templates/addedStudent.mustache
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <h1>Result of Adding Student:</h1>
+
+<p>Not implemented yet</p>
+
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
