+++ src/main/resources/templates/createresult.mustache
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <p>
+     This page has the result of creating a new alarm.
+ </p>
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
