+++ src/main/resources/templates/find.mustache
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <h2>Find a shift</h2>
+<p>Enter the key of the Shift you want to cover.</p>
+ <form action="/joinresult" method="get">
+    Key of your Shift:
+ <br>
+  <input type="text" name="key" value="jfu48rhgkbhdh">
+ <br>
+ <br>
+  <input type="submit" value="Join">
+ </form> 
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
