+++ src/main/resources/templates/home.mustache
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <h1>Welcome to ShiftCoverer!</h1>
+ <p>
+    An App that find your shift cover
+ </p>
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
