+++ src/main/resources/templates/joinresult.mustache
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <p>
+        This page has the result of joining an alarm.
+    </p>
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
